requirements.txt
streamlit>=1.31.0
openai>=1.12.0
pandas>=2.2.0
plotly>=5.18.0
google-cloud-bigquery>=3.14.1
google-auth>=2.27.0
db-dtypes>=1.2.0
python-dotenv>=1.0.0
openpyxl>=3.1.2  # Para soporte de archivos Excel
numpy>=1.26.0
altair>=5.2.0
supabase-py>=2.3.4 